requirements.txt
black==23.3.0
langchain==0.0.182
pinecone-client==2.2.1
pypdf==3.9.0y
tiktoken==0.4.0pip
unstructured==0.7.1
pdf2image
tabulate